fix(models): Change nutrition totals from Int to Double to match backend
Critical Schema Mismatch Fix

Root Cause:
Backend returns nutrition values as Doubles with decimals:
- protein_g total: 182.7 (not 182)
- carbs_g total: 345.5 (not 345)
- daily_breakdown protein_g: 120.7 (not 120)

But iOS models expected Ints:
- NutrientSummary.total: Int ❌
- DailyBreakdown.proteinG: Int ❌
- DailyBreakdown.carbsG: Int ❌
- DailyBreakdown.fatG: Int ❌

This caused JSON decoding error:
'Number 182.7 is not representable in Swift'

Solution:
Update iOS models to match backend schema:

NutrientSummary:
- total: Int → Double (backend calculates precise totals)

DailyBreakdown:
- proteinG: Int → Double (backend aggregates daily values)
- carbsG: Int → Double
- fatG: Int → Double

Note: calories stays as Int (backend returns integer calorie counts)

Impact:
- JSON decoding now succeeds
- Data is more accurate (precise decimal values preserved)
- Navigation works correctly for all date ranges
- Line charts display proper nutrition values

Testing:
- Navigate back through Apr dates
- Should see data load successfully
- Line charts should display with accurate values

+++ NomNom-iOS/NomNom/Core/Models/WeeklySummary.swift
@@ -31,7 +31,7 @@ struct WeeklySummaryResponse: Codable {
 }
 
 struct NutrientSummary: Codable {
-    let total: Int
+    let total: Double
     let average: Double
     let target: Int
     let percentage: Double?
@@ -40,9 +40,9 @@ struct NutrientSummary: Codable {
 struct DailyBreakdown: Codable {
     let date: String
     let calories: Int
-    let proteinG: Int
-    let carbsG: Int
-    let fatG: Int
+    let proteinG: Double
+    let carbsG: Double
+    let fatG: Double
 
     enum CodingKeys: String, CodingKey {
         case date
